Create first workspace based on client's rootUri
Currently the server will send a workspace/workspaceFolders request and
use the first folder that is reported as workspace root.  This doesn't
work with clients that don't support workspaceFolders capability.

The rootUri field of the response to 'initialize' is mandatory and can
be used to create the initial workspace too.  This patch does exactly
this.