Only call workspace/configuration when available
Not all clients implement the client capability: `configuration`, which
was added in version 3.6.0 of the Language Server Protocol. The LSP
Specification also states:

> A missing property should be interpreted as an absence of the capability.

Above claims are possible to verify by reading the mentioned spec. at:
https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#workspace_configuration

Hence this change modifies behaviour to only call the method on clients
explicitly announcing to support it.

This commit makes the lua-language-server work with vim-ale.

Thanks to Tom Lau for assistance in getting the test-setup updated.