Fix backwards compatability with `Lua.workspace.checkThirdParty`
I attempted to maintain backwards compatability in #2354 and #2406 but
didn't fully understand the config type system.

This approach uses `Type.Or` and I even tested it to confirm that it
works!

Thanks to @pysan3 for pointing out how a default would be used for
invalid types and to @sumneko for pointing out `Type.Or`.