Run diagnostics on unopened files when explicitly requested by the user
If I explicitly click on "Diagnose workspace" I would expect it to show
me *all* the diagnostics similar to how --check in a CI would check my
project.

This makes it easier for users to check their whole workspace after
making changes to some library or core function that potentially affects
a large number of files that are usually only checked once opened.