Send the correct `insertTextFormat`
Per LSP specification, `insertTextFormat` should be `2` for snippets and
`1` for plain text. Using `2` unconditionally can make clients behave in
unexpected ways.

The code calling into `provider.lua` sets `insertTextFormat` to 2 for
all snippets and leaves it nil for other types, so we can reuse that
information in `provider.lua` to make sure we send the correct value.